common & core & db: [#28] move DbCardRepository & DbDictionaryRepository from :common to :db-common; related moving and refactoring
