+++ common/build.gradle.kts
@@ -1,5 +1,3 @@
-@file:Suppress("UNUSED_VARIABLE")
-
 plugins {
     kotlin("multiplatform")
 }
@@ -14,6 +12,7 @@ kotlin {
         val kotlinDatetimeVersion: String by project
         val commonMain by getting {
             dependencies {
+                implementation(project(":db-common"))
                 implementation(kotlin("stdlib-common"))
                 api("org.jetbrains.kotlinx:kotlinx-datetime:$kotlinDatetimeVersion")
             }
