+++ db-common/src/main/kotlin/common/CommonMappers.kt
@@ -4,19 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.KotlinModule
-import com.gitlab.sszuev.flashcards.model.common.NONE
-import kotlinx.datetime.toJavaInstant
-import kotlinx.datetime.toKotlinInstant
-import java.time.temporal.ChronoUnit
 
-fun systemNow(): java.time.LocalDateTime =
-    java.time.OffsetDateTime.now(java.time.ZoneOffset.UTC).truncatedTo(ChronoUnit.MILLIS).toLocalDateTime()
-
-fun kotlinx.datetime.Instant?.asJava(): java.time.LocalDateTime =
-    (this?:kotlinx.datetime.Instant.NONE).toJavaInstant().atOffset(java.time.ZoneOffset.UTC).toLocalDateTime()
-
-fun java.time.LocalDateTime?.asKotlin(): kotlinx.datetime.Instant =
-    this?.toInstant(java.time.ZoneOffset.UTC)?.toKotlinInstant() ?: kotlinx.datetime.Instant.NONE
 
 private val mapper = ObjectMapper()
     .setSerializationInclusion(JsonInclude.Include.NON_EMPTY)
