+++ db-common/src/main/kotlin/repositories/DbCadStage.kt
@@ -1,0 +1,8 @@
+package com.gitlab.sszuev.flashcards.repositories
+
+enum class DbCadStage {
+    MOSAIC,
+    OPTIONS,
+    WRITING,
+    SELF_TEST,
+}
