+++ db-common/src/main/kotlin/repositories/DbDataException.kt
