+++ db-mem/src/main/kotlin/MemDbCardRepository.kt
@@ -1,11 +1,11 @@
 package com.gitlab.sszuev.flashcards.dbmem
 
-import com.gitlab.sszuev.flashcards.common.systemNow
 import com.gitlab.sszuev.flashcards.common.validateCardEntityForCreate
 import com.gitlab.sszuev.flashcards.common.validateCardEntityForUpdate
 import com.gitlab.sszuev.flashcards.repositories.DbCard
 import com.gitlab.sszuev.flashcards.repositories.DbCardRepository
 import com.gitlab.sszuev.flashcards.repositories.DbDataException
+import com.gitlab.sszuev.flashcards.systemNow
 
 class MemDbCardRepository(
     dbConfig: MemDbConfig = MemDbConfig(),
