+++ db-mem/src/main/kotlin/MemDbDictionaryRepository.kt
@@ -1,9 +1,9 @@
 package com.gitlab.sszuev.flashcards.dbmem
 
-import com.gitlab.sszuev.flashcards.common.systemNow
 import com.gitlab.sszuev.flashcards.repositories.DbDataException
 import com.gitlab.sszuev.flashcards.repositories.DbDictionary
 import com.gitlab.sszuev.flashcards.repositories.DbDictionaryRepository
+import com.gitlab.sszuev.flashcards.systemNow
 
 class MemDbDictionaryRepository(
     dbConfig: MemDbConfig = MemDbConfig(),
