+++ db-mem/src/test/kotlin/MemDatabaseTest.kt
@@ -1,11 +1,11 @@
 package com.gitlab.sszuev.flashcards.dbmem
 
-import com.gitlab.sszuev.flashcards.common.systemNow
 import com.gitlab.sszuev.flashcards.dbmem.dao.MemDbCard
 import com.gitlab.sszuev.flashcards.dbmem.dao.MemDbUser
 import com.gitlab.sszuev.flashcards.dbmem.dao.MemDbWord
 import com.gitlab.sszuev.flashcards.dbmem.testutils.classPathResourceDir
 import com.gitlab.sszuev.flashcards.dbmem.testutils.copyClassPathDataToDir
+import com.gitlab.sszuev.flashcards.systemNow
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Order
 import org.junit.jupiter.api.Test
